+++ spec/unit/resources/xdebug_spec.rb
@@ -24,7 +24,8 @@ describe 'codenamephp_php_xdebug' do
       is_expected.to enable_codenamephp_php_config('xdebug').with(
         cookbook_file: 'xdebug.ini',
         config_name: 'xdebug-custom.ini',
-        php_versions: %w[99 88]
+        php_versions: %w[99 88],
+        services: %w[cli apache2 fpm]
       )
     }
   end
@@ -43,7 +44,40 @@ describe 'codenamephp_php_xdebug' do
       is_expected.to delete_codenamephp_php_config('xdebug').with(
         cookbook_file: 'xdebug.ini',
         config_name: 'xdebug-custom.ini',
-        php_versions: %w[99 88]
+        php_versions: %w[99 88],
+        services: %w[cli apache2 fpm]
+      )
+    }
+  end
+
+  context 'Install with services' do
+    recipe do
+      codenamephp_php_xdebug 'install xdebug' do
+        php_versions %w[99 88]
+        action :install
+        services %w[test1 test2]
+      end
+    end
+
+    it {
+      is_expected.to enable_codenamephp_php_config('xdebug').with(
+        services: %w[test1 test2]
+      )
+    }
+  end
+
+  context 'Remove with services' do
+    recipe do
+      codenamephp_php_xdebug 'install xdebug' do
+        php_versions %w[99 88]
+        action :remove
+        services %w[test1 test2]
+      end
+    end
+
+    it {
+      is_expected.to delete_codenamephp_php_config('xdebug').with(
+        services: %w[test1 test2]
       )
     }
   end