+++ test/fixtures/cookbooks/test/metadata.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+name 'test'
+maintainer 'Bastian Schwarz'
+maintainer_email 'bastian@codename-php.de'
+license 'Apache-2.0'
+description 'Cookbook to install php'
+version '1.0.0'
+
+supports 'debian', '~>9.1'
+
+depends 'codenamephp_php'